Move purchaseable aug files around
